ETL/requirements.txt
psycopg2-binary==2.9.3
elasticsearch==8.0.0
tenacity==8.0.1




